+++ pybamm/models/full_battery_models/lithium_ion/spme.py
@@ -61,13 +61,11 @@ class SPMe(SPM):
         self.submodels[
             "electrolyte transport efficiency"
         ] = pybamm.transport_efficiency.Bruggeman(
-            self.param, "Electrolyte", self.options, True
+            self.param, "Electrolyte", self.options
         )
         self.submodels[
             "electrode transport efficiency"
-        ] = pybamm.transport_efficiency.Bruggeman(
-            self.param, "Electrode", self.options, True
-        )
+        ] = pybamm.transport_efficiency.Bruggeman(self.param, "Electrode", self.options)
 
     def set_solid_submodel(self):
         for domain in ["negative", "positive"]: