+++ pybamm/models/submodels/electrolyte_conductivity/integrated_conductivity.py
@@ -42,9 +42,7 @@ class Integrated(BaseElectrolyteConductivity):
         param = self.param
         c_e_av = variables["X-averaged electrolyte concentration [mol.m-3]"]
 
-        i_boundary_cc_0 = variables[
-            "Leading-order current collector current density [A.m-2]"
-        ]
+        i_boundary_cc = variables["Current collector current density [A.m-2]"]
         c_e_n = variables["Negative electrolyte concentration [mol.m-3]"]
         c_e_s = variables["Separator electrolyte concentration [mol.m-3]"]
         c_e_p = variables["Positive electrolyte concentration [mol.m-3]"]
@@ -84,14 +82,14 @@ class Integrated(BaseElectrolyteConductivity):
         chi_av_p = pybamm.PrimaryBroadcast(chi_av, "positive electrode")
 
         # electrolyte current
-        i_e_n = i_boundary_cc_0 * x_n / L_n
-        i_e_s = pybamm.PrimaryBroadcast(i_boundary_cc_0, "separator")
-        i_e_p = i_boundary_cc_0 * (1 - x_p) / L_p
+        i_e_n = i_boundary_cc * x_n / L_n
+        i_e_s = pybamm.PrimaryBroadcast(i_boundary_cc, "separator")
+        i_e_p = i_boundary_cc * (1 - x_p) / L_p
         i_e = pybamm.concatenation(i_e_n, i_e_s, i_e_p)
 
-        i_e_n_edge = i_boundary_cc_0 * x_n_edge / L_n
-        i_e_s_edge = pybamm.PrimaryBroadcastToEdges(i_boundary_cc_0, "separator")
-        i_e_p_edge = i_boundary_cc_0 * (1 - x_p_edge) / L_p
+        i_e_n_edge = i_boundary_cc * x_n_edge / L_n
+        i_e_s_edge = pybamm.PrimaryBroadcastToEdges(i_boundary_cc, "separator")
+        i_e_p_edge = i_boundary_cc * (1 - x_p_edge) / L_p
 
         # electrolyte potential
         indef_integral_n = pybamm.IndefiniteIntegral(