+++ pybamm/models/submodels/electrolyte_diffusion/__init__.py
@@ -1,6 +1,4 @@
 from .base_electrolyte_diffusion import BaseElectrolyteDiffusion
 from .leading_order_diffusion import LeadingOrder
-from .first_order_diffusion import FirstOrder
-from .composite_diffusion import Composite
 from .full_diffusion import Full
 from .constant_concentration import ConstantConcentration