+++ pybamm/models/submodels/interface/base_interface.py
@@ -379,19 +379,6 @@ class BaseInterface(pybamm.BaseSubModel):
 
         return variables
 
-    def _get_standard_surface_potential_difference_variables(self, delta_phi):
-        domain, Domain = self.domain_Domain
-
-        # Broadcast if necessary
-        if delta_phi.domain == ["current collector"]:
-            delta_phi = pybamm.PrimaryBroadcast(delta_phi, f"{domain} electrode")
-
-        variables = {
-            f"{Domain} electrode surface potential difference [V]": delta_phi,
-        }
-
-        return variables
-
     def _get_standard_size_distribution_interfacial_current_variables(self, j):
         """
         Interfacial current density variables that depend on particle size R,